--- a/package.json
+++ b/package.json
@@ -32,11 +32,9 @@
     "@typescript-eslint/parser": "^5.30.7",
     "fastify": "^4.0.1",
     "jsdoc": "^3.6.10",
-    "simple-get": "^4.0.1",
     "standard": "^17.0.0",
     "tap": "^16.3.0",
-    "tsd": "^0.22.0",
-    "typescript": "^4.7.4"
+    "tsd": "^0.22.0"
   },
   "standard": {
     "ignore": [