--- a/test/favicon.test.js
+++ b/test/favicon.test.js
@@ -16,102 +16,66 @@
 'use strict'
 
 const test = require('tap').test
-const sget = require('simple-get').concat
 const Fastify = require('fastify')
 
-test('default favicon does not return an error, but a good response (200) and some content', (t) => {
-  // t.plan(6)
+test('default favicon does not return an error, but a good response (200) and some content', async (t) => {
+  t.plan(5)
+
   const defaultPath = './src'
   const fastify = Fastify()
   t.teardown(() => { fastify.close() })
 
-  fastify.register(require('../')) // configure this plugin with its default options
+  fastify.register(require('..'), {
+    path: defaultPath,
+    maxAge: 1337
+  }) // configure this plugin with its default options
 
-  fastify.listen({ port: 0 }, (err, address) => {
-    t.error(err)
+  await fastify.listen({ port: 0 })
 
-    sget({
-      method: 'GET',
-      timeout: 2000,
-      url: `${address}/favicon.ico`
-    }, (err, response, body) => {
-      t.error(err)
-      t.equal(response.statusCode, 200)
-      t.equal(response.headers['content-type'], 'image/x-icon')
-      // add check on file contents, or at least file size ...
-      const fs = require('fs')
-      const contents = fs.readFileSync(`${defaultPath}/favicon.ico`)
-      // optional, add some assertions with standard Node.js assert statements, as a sample
-      const assert = require('assert')
-      assert(contents !== null)
-      t.ok(contents)
-      t.strictSame(contents.length, body.length)
-      t.end()
-    })
+  const response = await fastify.inject({
+    method: 'GET',
+    timeout: 2000,
+    url: '/favicon.ico'
   })
+
+  t.equal(response.statusCode, 200)
+  t.equal(response.headers['cache-control'], 'max-age=1337')
+  t.equal(response.headers['content-type'], 'image/x-icon')
+
+  // add check on file contents, or at least file size ...
+  const fs = require('fs')
+  const contents = fs.readFileSync(`${defaultPath}/favicon.ico`)
+  t.ok(contents)
+  t.strictSame(contents.length, response.body.length)
 })
 
-test('return a favicon configured in a custom path', (t) => {
+test('return a favicon configured in a custom path', async (t) => {
+  t.plan(5)
+
   const pathSample = './test'
   const fastify = Fastify()
+
   t.teardown(() => { fastify.close() })
-  fastify.register(require('../'), {
+
+  await fastify.register(require('../'), {
     path: pathSample
   })
 
-  fastify.listen({ port: 0 }, (err, address) => {
-    t.error(err)
+  await fastify.listen({ port: 0 })
 
-    sget({
-      method: 'GET',
-      timeout: 2000,
-      url: `${address}/favicon.ico`
-    }, (err, response, body) => {
-      t.error(err)
-      t.equal(response.statusCode, 200)
-      t.equal(response.headers['content-type'], 'image/x-icon')
-      // add check on file contents, or at least file size ...
-      const fs = require('fs')
-      const contents = fs.readFileSync(`${pathSample}/favicon.ico`)
-      // optional, add some assertions with standard Node.js assert statements, as a sample
-      const assert = require('assert')
-      assert(contents !== null)
-      t.ok(contents)
-      t.strictSame(contents.length, body.length)
-      t.end()
-    })
-  })
-})
-
-test('return default favicon because that in the custom path is not found', (t) => {
-  const pathSample = './test/not-existing-img-path' // path that here does not exist, good for this test
-  const defaultPath = './src'
-  const fastify = Fastify()
-  t.teardown(() => { fastify.close() })
-  fastify.register(require('../'), {
-    path: pathSample
+  const response = await fastify.inject({
+    method: 'GET',
+    timeout: 2000,
+    url: '/favicon.ico'
   })
 
-  fastify.listen({ port: 0 }, (err, address) => {
-    t.error(err)
+  t.equal(response.statusCode, 200)
+  t.equal(response.headers['cache-control'], 'max-age=86400')
+  t.equal(response.headers['content-type'], 'image/x-icon')
 
-    sget({
-      method: 'GET',
-      timeout: 2000,
-      url: `${address}/favicon.ico`
-    }, (err, response, body) => {
-      t.error(err)
-      t.equal(response.statusCode, 200)
-      t.equal(response.headers['content-type'], 'image/x-icon')
-      // add check on file contents, or at least file size ...
-      const fs = require('fs')
-      const contents = fs.readFileSync(`${defaultPath}/favicon.ico`)
-      // optional, add some assertions with standard Node.js assert statements, as a sample
-      const assert = require('assert')
-      assert(contents !== null)
-      t.ok(contents)
-      t.strictSame(contents.length, body.length)
-      t.end()
-    })
-  })
+  // add check on file contents, or at least file size ...
+  const fs = require('fs')
+  const contents = fs.readFileSync(`${pathSample}/favicon.ico`)
+  t.ok(contents)
+  t.strictSame(contents.length, response.body.length)
 })