It is not possible anymore to change the favicon on the fly. It caches it once and then serves it all the time. The server start is interrupted if it could not load the favicon. Adds caching to the favicon to reduce the pressure on the server.

before:
```sh
uzlopak@Battlestation:~/workspace/fastify-favicon$ autocannon http://localhost:3000/favicon.ico
Running 10s test @ http://localhost:3000/favicon.ico
10 connections


┌─────────┬──────┬───────┬───────┬───────┬──────────┬─────────┬───────┐
│ Stat    │ 2.5% │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev   │ Max   │
├─────────┼──────┼───────┼───────┼───────┼──────────┼─────────┼───────┤
│ Latency │ 1 ms │ 44 ms │ 50 ms │ 51 ms │ 42.34 ms │ 9.71 ms │ 55 ms │
└─────────┴──────┴───────┴───────┴───────┴──────────┴─────────┴───────┘
┌───────────┬────────┬────────┬────────┬────────┬────────┬─────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%    │ 97.5%  │ Avg    │ Stdev   │ Min    │
├───────────┼────────┼────────┼────────┼────────┼────────┼─────────┼────────┤
│ Req/Sec   │ 215    │ 215    │ 222    │ 296    │ 232.9  │ 24.28   │ 215    │
├───────────┼────────┼────────┼────────┼────────┼────────┼─────────┼────────┤
│ Bytes/Sec │ 271 kB │ 271 kB │ 279 kB │ 372 kB │ 293 kB │ 30.5 kB │ 270 kB │
└───────────┴────────┴────────┴────────┴────────┴────────┴─────────┴────────┘

Req/Bytes counts sampled once per second.
# of samples: 10

2k requests in 10.02s, 2.93 MB read
```

after:
```sh
uzlopak@Battlestation:~/workspace/fastify-favicon$ autocannon http://localhost:3000/favicon.ico
Running 10s test @ http://localhost:3000/favicon.ico
10 connections


┌─────────┬──────┬──────┬───────┬──────┬─────────┬─────────┬───────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%  │ Avg     │ Stdev   │ Max   │
├─────────┼──────┼──────┼───────┼──────┼─────────┼─────────┼───────┤
│ Latency │ 0 ms │ 0 ms │ 0 ms  │ 0 ms │ 0.01 ms │ 0.06 ms │ 11 ms │
└─────────┴──────┴──────┴───────┴──────┴─────────┴─────────┴───────┘
┌───────────┬───────┬───────┬─────────┬─────────┬──────────┬─────────┬───────┐
│ Stat      │ 1%    │ 2.5%  │ 50%     │ 97.5%   │ Avg      │ Stdev   │ Min   │
├───────────┼───────┼───────┼─────────┼─────────┼──────────┼─────────┼───────┤
│ Req/Sec   │ 45695 │ 45695 │ 62687   │ 66687   │ 61186.91 │ 5372.86 │ 45678 │
├───────────┼───────┼───────┼─────────┼─────────┼──────────┼─────────┼───────┤
│ Bytes/Sec │ 58 MB │ 58 MB │ 79.6 MB │ 84.7 MB │ 77.7 MB  │ 6.82 MB │ 58 MB │
└───────────┴───────┴───────┴─────────┴─────────┴──────────┴─────────┴───────┘

Req/Bytes counts sampled once per second.
# of samples: 11

673k requests in 11.01s, 855 MB read
```