@smartiniOnGitHub 2022-08-25
Hi @Uzlopak thanks for this PR, it's really a big improvement and cleanup of some code.
Now I'll merge and publish a new release ASAP (max in a few days). 
Thanks a lot :)